Applied cube controller to cube page
$routeProvider's resolve lets you specify dependencies that have to be loaded before you will render the page. We are waiting on scripts to load before rendering the page. Here, that means that the page is guaranteed to be functional before we render it. If 3js took 3 seconds to render the cube onto the canvas, then we would see a page with just our html on it, buttons on top of nothing.
If we resolve the promise after the scripts load, then the page won't be rendered for 3 seconds, but it will have your JS on it- the canvas will be there.

The common practice with angular is to have a single module per page. I was trying to use different modules for the home and cube controllers. That doesn't make sense because angular only compiles the module object once, on DOMContentLoaded.

Instead, I ended up defining both the home and cube controllers on the angular module, before bootstrapping the app. That way, when $apply gets called in the resolve for routeProvider, my model already be defined, and the view will get updated appropriately.